refactor: split Profile to Form and Nav components and change some imports
