+++ src/components/Modal/IngredientDetails/index.js
@@ -1,0 +1,1 @@
+export { default } from "./IngredientDetails";
