+++ src/components/Profile/ProfileNav/ProfileNav.module.css
@@ -1,0 +1,17 @@
+.menu {
+  width: 320px;
+}
+
+.list {
+  list-style: none;
+  margin: 0;
+  padding: 0;
+}
+
+.link {
+  text-decoration: none;
+}
+
+.text_active {
+  color: #f2f2f3;
+}
