+++ src/components/Profile/ProfileNav/index.js
@@ -1,0 +1,1 @@
+export { default } from "./ProfileNav";
