+++ src/pages/index.js
@@ -1,0 +1,25 @@
+import LoginPage from "./login";
+import RegisterPage from "./register";
+import ForgotPasswordPage from "./forgot-password";
+import ResetPasswordPage from "./reset-password";
+import ProfilePage from "./profile";
+import NotFound404Page from "./not-found-404";
+import IngredientPage from "./ingredient";
+import FeedPage from "./feed";
+import FeedOrderPage from "./feed-order";
+import ProfileOrdersPage from "./profile-orders";
+import ProfileOrderInfoPage from "./profile-order-info";
+
+export {
+  LoginPage,
+  RegisterPage,
+  ForgotPasswordPage,
+  ResetPasswordPage,
+  ProfilePage,
+  NotFound404Page,
+  IngredientPage,
+  FeedPage,
+  FeedOrderPage,
+  ProfileOrdersPage,
+  ProfileOrderInfoPage,
+};
