--- src/pages/profile.js
@@ -1,9 +1,0 @@
-import React from "react";
-
-import Profile from "../components/Profile/Profile";
-
-function ProfilePage() {
-  return <Profile />;
-}
-
-export default ProfilePage;
